fclShopMisc FUN_00400740 nd80->8 (alias routing), FUN_00400A90 nd77->75; mt_evtMisc final 43/17

+++ src/Yajima/y_misc.c
@@ -6240,7 +6240,7 @@ void FUN_0042cd80(int param_1)
   s16 *offsetX;
   s16 *offsetY;
   u32 random;
-  s32 count;
+  s8 count;
   s32 outer;
   s32 inner;
   s32 end;
